refactor: move TestMinutesInRadians to internal clockface_test.go for package visibility
