--- maths/clockface/clockface_acceptance_test.go
@@ -1,126 +1,0 @@
-package clockface_test
-
-import (
-	"bytes"
-	"encoding/xml"
-	"math"
-	"testing"
-	"time"
-
-	"example.com/hello/maths/clockface"
-)
-
-type SVG struct {
-	XMLName xml.Name `xml:"svg"`
-	Text    string   `xml:",chardata"`
-	Xmlns   string   `xml:"xmlns,attr"`
-	Width   string   `xml:"width,attr"`
-	Height  string   `xml:"height,attr"`
-	ViewBox string   `xml:"viewBox,attr"`
-	Version string   `xml:"version,attr"`
-	Circle  Circle   `xml:"circle"`
-	Line    []Line   `xml:"line"`
-}
-
-type Line struct {
-	X1 float64 `xml:"x1,attr"`
-	Y1 float64 `xml:"y1,attr"`
-	X2 float64 `xml:"x2,attr"`
-	Y2 float64 `xml:"y2,attr"`
-}
-
-type Circle struct {
-	Cx float64 `xml:"cx,attr"`
-	Cy float64 `xml:"cy,attr"`
-	R  float64 `xml:"r,attr"`
-}
-
-func TestSVGWriterSecondHand(t *testing.T) {
-	cases := []struct {
-		time time.Time
-		line Line
-	}{
-		{
-			simpleTime(0, 0, 0),
-			Line{150, 150, 150, 60},
-		},
-		{
-			simpleTime(0, 0, 30),
-			Line{150, 150, 150, 240},
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(testName(c.time), func(t *testing.T) {
-			b := bytes.Buffer{}
-			clockface.SVGWriter(&b, c.time)
-
-			svg := SVG{}
-			xml.Unmarshal(b.Bytes(), &svg)
-
-			if !containsLine(c.line, svg.Line) {
-				t.Errorf("Expected to find the second hand line %+v, in the SVG lines %+v", c.line, svg.Line)
-			}
-		})
-	}
-}
-
-func containsLine(l Line, ls []Line) bool {
-	for _, line := range ls {
-		if line == l {
-			return true
-		}
-	}
-	return false
-}
-
-func simpleTime(hours, minutes, seconds int) time.Time {
-	return time.Date(312, time.October, 28, hours, minutes, seconds, 0, time.UTC)
-}
-
-func testName(t time.Time) string {
-	return t.Format("15:04:05")
-}
-
-func TestSVGWriterMinuteHand(t *testing.T) {
-	cases := []struct {
-		time time.Time
-		line Line
-	}{
-		{
-			simpleTime(0, 0, 0),
-			Line{150, 150, 150, 70},
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(testName(c.time), func(t *testing.T) {
-			b := bytes.Buffer{}
-			clockface.SVGWriter(&b, c.time)
-
-			svg := SVG{}
-			xml.Unmarshal(b.Bytes(), &svg)
-
-			if !containsLine(c.line, svg.Line) {
-				t.Errorf("Expected to find the minute hand line %+v, in the SVG lines %+v", c.line, svg.Line)
-			}
-		})
-	}
-}
-
-func TestMinutesInRadians(t *testing.T) {
-	cases := []struct {
-		time  time.Time
-		angle float64
-	}{
-		{simpleTime(0, 0, 30), math.Pi},
-	}
-	for _, c := range cases {
-		t.Run(testName(c.time), func(t *testing.T) {
-			got := minutesInRadians(c.time)
-			if got != c.angle {
-				t.Fatalf("Wanted %v radians, but got %v", c.angle, got)
-			}
-		})
-	}
-}
